+++ src/artm/regularizer/topic_segmentation_ptdw.cc
@@ -1,6 +1,6 @@
 // Copyright 2017, Additive Regularization of Topic Models.
 
-// Author: Anastasia Bayandina
+// Author: Anastasia Bayandina (anast.bayandina@gmail.com)
 
 #include <vector>
 #include <algorithm>
@@ -18,7 +18,7 @@ void TopicSegmentationPtdwAgent::Apply(int item_index, int inner_iter,
                                        ::artm::utility::LocalPhiMatrix<float>* ptdw) const {
   int local_token_size = ptdw->no_rows();
   int num_topics = ptdw->no_columns();
-  std::vector<double> background_probability(local_token_size, 0.0);
+  std::vector<float> background_probability(local_token_size, 0.0f);
   // if background topics are given, count probability for each word to be background
   if (config_.background_topic_names().size()) {
     std::vector<bool> is_background_topic = core::is_member(args_.topic_name(), config_.background_topic_names());
@@ -33,12 +33,12 @@ void TopicSegmentationPtdwAgent::Apply(int item_index, int inner_iter,
   }
 
   int h = config_.window();
-  double threshold_topic_change = config_.threshold();
+  float threshold_topic_change = config_.threshold();
   ::artm::utility::DenseMatrix<float> copy_ptdw(*ptdw);
-  std::vector<double> left_distribution(num_topics, 0.0);
-  std::vector<double> right_distribution(num_topics, 0.0);
-  double left_weights = 0.0;
-  double right_weights = 0.0;
+  std::vector<float> left_distribution(num_topics, 0.0f);
+  std::vector<float> right_distribution(num_topics, 0.0f);
+  float left_weights = 0.0f;
+  float right_weights = 0.0f;
   int l_topic, r_topic;  // topic ids on which maximum of the distribution is reached
   bool changes_topic = false;
   int main_topic = 0;
@@ -88,10 +88,10 @@ void TopicSegmentationPtdwAgent::Apply(int item_index, int inner_iter,
     auto re = right_distribution.end();
     l_topic = std::distance(lb, std::max_element(lb, le));
     r_topic = std::distance(rb, std::max_element(rb, re));
-    double ll =  left_distribution[l_topic];
-    double rl = right_distribution[l_topic];
-    double rr = right_distribution[r_topic];
-    double lr =  left_distribution[r_topic];
+    float ll =  left_distribution[l_topic];
+    float rl = right_distribution[l_topic];
+    float rr = right_distribution[r_topic];
+    float lr =  left_distribution[r_topic];
     changes_topic = ((ll / left_weights  - rl / right_weights) / 2 +
                      (rr / right_weights - lr / left_weights)  / 2 > threshold_topic_change);
     if (changes_topic) {
@@ -109,7 +109,7 @@ void TopicSegmentationPtdwAgent::Apply(int item_index, int inner_iter,
 
 std::shared_ptr<RegularizePtdwAgent>
 TopicSegmentationPtdw::CreateRegularizePtdwAgent(const Batch& batch,
-                                                 const ProcessBatchesArgs& args, double tau) {
+                                                 const ProcessBatchesArgs& args, float tau) {
   TopicSegmentationPtdwAgent* agent = new TopicSegmentationPtdwAgent(config_, args, tau);
   std::shared_ptr<RegularizePtdwAgent> retval(agent);
   return retval;
