+++ src/artm/regularizer/topic_segmentation_ptdw.h
@@ -23,7 +23,7 @@ class TopicSegmentationPtdwAgent : public RegularizePtdwAgent {
   ProcessBatchesArgs args_;
 
  public:
-  TopicSegmentationPtdwAgent(const TopicSegmentationPtdwConfig& config, const ProcessBatchesArgs& args, double tau)
+  TopicSegmentationPtdwAgent(const TopicSegmentationPtdwConfig& config, const ProcessBatchesArgs& args, float tau)
     : config_(config), args_(args) {}
 
   virtual void Apply(int item_index, int inner_iter, ::artm::utility::LocalPhiMatrix<float>* ptdw) const;
@@ -35,7 +35,7 @@ class TopicSegmentationPtdw : public RegularizerInterface {
   : config_(config) {}
 
   virtual std::shared_ptr<RegularizePtdwAgent>
-  CreateRegularizePtdwAgent(const Batch& batch, const ProcessBatchesArgs& args, double tau);
+  CreateRegularizePtdwAgent(const Batch& batch, const ProcessBatchesArgs& args, float tau);
 
   virtual bool Reconfigure(const RegularizerConfig& config);
